Fix unsigned comparisons and assignment order in evt_case, unit_object_switch, unit_object_tree
evt_case.c: Move evtCode/priority before memcpy, change activeConditionId
  param from u16 to s32 to match target register allocation patterns
unit_object_switch.c: Cast userdata[0] to u32 for unsigned comparison
unit_object_tree.c: Same u32 cast for _rumble_tree

+++ src/unit/unit_object_switch.c
@@ -130,7 +130,7 @@ USER_FUNC(_rumble_switch) {
         obj->mPosition.y += delta;
         event->userdata[0]++;
     }
-    if (event->userdata[0] >= 0x22) {
+    if ((u32)event->userdata[0] >= 0x22) {
         return EVT_RETURN_DONE;
     }
     return 0;
